+++ src/project.ts
@@ -1,6 +1,0 @@
-export interface Project {
-    window?: any
-    workspace?: any
-    rootPath?: string
-    extensionPath?: string
-}