+++ src/tester.ts
@@ -1,37 +1,47 @@
 import { Disposable, TextDocument, TextEditor } from 'vscode'
+import { Phpunit, State } from './phpunit'
 
 import { DecorateManager } from './decorate-manager'
 import { DiagnosticManager } from './diagnostic-manager'
-import { Phpunit } from './phpunit'
-import { Project } from './project'
 import { Store } from './store'
+
+export interface Project {
+    window?: any
+    workspace?: any
+    rootPath?: string
+    extensionPath?: string
+}
 
 export class Tester {
     private disposable: Disposable
+    private window: any
+    private workspace: any
 
     constructor(
-        private project: Project,
+        project: Project,
         private phpunit: Phpunit,
         private decorateManager: DecorateManager,
         private diagnosticManager: DiagnosticManager,
         private store: Store = new Store()
-    ) {}
+    ) {
+        this.window = project.window
+        this.workspace = project.workspace
+    }
 
     subscribe(): this {
         const subscriptions: Disposable[] = []
-        const { window, workspace } = this.project
 
-        // workspace.onDidOpenTextDocument(this.restore.bind(this), null, subscriptions)
-        workspace.onWillSaveTextDocument(this.restore.bind(this), null, subscriptions)
-        // workspace.onDidSaveTextDocument(this.restore.bind(this), null, subscriptions)
-        // workspace.onDidChangeTextDocument((document: TextDocument) => {
+        // this.workspace.onDidOpenTextDocument(this.restore.bind(this), null, subscriptions)
+        this.workspace.onWillSaveTextDocument(this.exec.bind(this), null, subscriptions)
+        // this.workspace.onDidSaveTextDocument(this.restore.bind(this), null, subscriptions)
+        // this.workspace.onDidChangeTextDocument((document: TextDocument) => {
         //     const editor = this.getActiveTextEditor()
         //     if (editor && document === editor.document) {
         //         this.restore()
         //     }
         // }, null, subscriptions)
-        // window.onDidChangeActiveTextEditor(this.restore.bind(this), null, subscriptions)
-        window.onDidChangeActiveTextEditor(this.restore.bind(this), null, subscriptions)
+        // this.window.onDidChangeActiveTextEditor(this.restore.bind(this), null, subscriptions)
+        this.window.onDidChangeActiveTextEditor(this.restore.bind(this), null, subscriptions)
 
         this.restore()
 
@@ -41,22 +51,36 @@ export class Tester {
     }
 
     async exec() {
-        if (this.isExecutable() === false) {
-            return
+        this.clearDecoratedGutter()
+
+        try {
+            if (!this.editor || !this.document) {
+                return
+            }
+
+            const fileName = this.document.fileName
+            const messages = await this.phpunit.exec(fileName, this.document.getText())
+            this.store.put(messages)
+
+            this.decoratedGutter()
+            this.handleDiagnostic()
+        } catch (e) {
+            switch (e) {
+                case State.PHPUNIT_NOT_FOUND:
+                    this.window.showErrorMessage('composer require phpunit/phpunit')
+                    console.error(this.phpunit.getLastOutput())
+                    break
+                case State.PHPUNIT_EXECUTE_ERROR:
+                    this.window.showErrorMessage('composer install')
+                    console.error(this.phpunit.getLastOutput())
+                    break
+                case State.NOT_RUNNABLE:
+                    break
+            }
         }
-
-        const messages = await this.phpunit.exec(this.document.fileName)
-        this.store.put(messages)
-
-        this.decoratedGutter()
-        this.handleDiagnostic()
     }
 
     restore(): void {
-        if (this.isExecutable() === false) {
-            return
-        }
-
         if (this.store.has(this.document.fileName)) {
             this.decoratedGutter()
             this.handleDiagnostic()
@@ -77,36 +101,19 @@ export class Tester {
         this.decorateManager.decoratedGutter(this.store, this.editor)
     }
 
+    private clearDecoratedGutter() {
+        this.decorateManager.clearDecoratedGutter(this.editor)
+    }
+
     private handleDiagnostic() {
         this.diagnosticManager.handle(this.store, this.editor)
     }
 
-    private isExecutable() {
-        const keywords = new RegExp(
-            [
-                'PHPUnit\\\\Framework\\\\TestCase',
-                'PHPUnit\\Framework\\TestCase',
-                'PHPUnit_Framework_TestCase',
-                'TestCase',
-            ].join('|')
-        )
-
-        if (!this.editor || !this.document || keywords.test(this.document.getText()) === false) {
-            return false
-        }
-
-        if (/\.git\.php$/.test(this.document.fileName) === true) {
-            return false
-        }
-
-        return true
-    }
-
     get editor(): TextEditor {
-        return this.project.window.activeTextEditor
+        return this.window.activeTextEditor
     }
 
     get document(): TextDocument {
-        return this.editor.document
+        return this.window.activeTextEditor.document
     }
 }